alx_travel_app/alx_travel_app/requirement.txt
amqp==5.3.1
asgiref==3.9.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
cffi==1.17.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.4
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
isort==6.0.1
kombu==5.5.4
mypy==1.16.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
pycparser==2.22
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rabbitmq==0.2.0
ruff==0.12.2
six==1.17.0
sqlparse==0.5.3
ty==0.0.1a13
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
